fixed according to the comments in file review-2021-03-07.pdf

+++ src/lib/DemodulateLibMath.cpp
@@ -1,8 +1,9 @@
 #include "DemodulateLibMath.h"
+#include <cmath>
 
 void DemodulateAm(Flvec & output,const  IQvec & source)
 {
-	for (int i = 0; i != source.size(); ++i)
+	for (size_t i = 0; i != source.size(); ++i)
 	{
 		output[i] = sqrtf(source[i].I * source[i].I + source[i].Q * source[i].Q);
 	}
@@ -10,7 +11,7 @@ void DemodulateAm(Flvec & output,const  IQvec & source)
 
 void DemodulateFm(Flvec & output,const IQvec & source)
 {
-	for (int i = 0; i != source.size() - 1; ++i)
+	for (size_t i = 0; i != source.size() - 1; ++i)
 	{
 		output[i] = RotateElement(source[i + 1], source[i].GetConjugated()).getPhase();
 	}
@@ -18,7 +19,7 @@ void DemodulateFm(Flvec & output,const IQvec & source)
 
 IQElement RotateElement(const IQElement & a, const IQElement & b)
 {
-	return IQElement(a.I * b.I - a.Q * b.Q,a.Q * b.I + a.I * b.Q);
+	return IQElement(a.I * b.I - a.Q * b.Q, a.Q * b.I + a.I * b.Q);
 }
 
 void ApplyFilter(IQvec & src, const float * filter, const int & size)
@@ -39,4 +40,3 @@ void ApplyFilter(IQvec & src, const float * filter, const int & size)
 		src[i] = e;
 	}
 }
-
